+++ rocprim/include/rocprim/device/device_scan.hpp
@@ -202,7 +202,7 @@ auto scan_impl(void * temporary_storage,
 
     constexpr unsigned int block_size = config::block_size;
     constexpr unsigned int items_per_thread = config::items_per_thread;
-    constexpr auto items_per_block = block_size * items_per_thread;
+    constexpr auto items_per_block = (size_t)block_size * items_per_thread;
 
     // Calculate required temporary storage
     if(temporary_storage == nullptr)
